// frontend/src/autogen/proto.d.ts
export interface IArrow {
  /** Arrow data */
  data?: Uint8Array | null

  /** Arrow styler */
  styler?: IStyler | null

  /** Arrow width */
  width?: number | null

  /** Arrow height */
  height?: number | null

  /** Arrow useContainerWidth */
  useContainerWidth?: boolean | null

  /** Arrow id */
  id?: string | null

  /** Arrow columns */
  columns?: string | null

  /** Arrow editingMode */
  editingMode?: Arrow.EditingMode | null

  /** Arrow disabled */
  disabled?: boolean | null

  /** Arrow formId */
  formId?: string | null
}

/** Represents an Arrow. */
export class Arrow implements IArrow {
  /**
   * Constructs a new Arrow.
   * @param [properties] Properties to set
   */
  constructor(properties?: IArrow)

  /** Arrow data. */
  public data: Uint8Array

  /** Arrow styler. */
  public styler?: IStyler | null

  /** Arrow width. */
  public width: number

  /** Arrow height. */
  public height: number

  /** Arrow useContainerWidth. */
  public useContainerWidth: boolean

  /** Arrow id. */
  public id: string

  /** Arrow columns. */
  public columns: string

  /** Arrow editingMode. */
  public editingMode: Arrow.EditingMode

  /** Arrow disabled. */
  public disabled: boolean

  /** Arrow formId. */
  public formId: string

  /**
   * Creates a new Arrow instance using the specified properties.
   * @param [properties] Properties to set
   * @returns Arrow instance
   */
  public static create(properties?: IArrow): Arrow

  /**
   * Encodes the specified Arrow message. Does not implicitly {@link Arrow.verify|verify} messages.
   * @param message Arrow message or plain object to encode
   * @param [writer] Writer to encode to
   * @returns Writer
   */
  public static encode(
    message: IArrow,
    writer?: $protobuf.Writer
  ): $protobuf.Writer

  /**
   * Encodes the specified Arrow message, length delimited. Does not implicitly {@link Arrow.verify|verify} messages.
   * @param message Arrow message or plain object to encode
   * @param [writer] Writer to encode to
   * @returns Writer
   */
  public static encodeDelimited(
    message: IArrow,
    writer?: $protobuf.Writer
  ): $protobuf.Writer

  /**
   * Decodes an Arrow message from the specified reader or buffer.
   * @param reader Reader or buffer to decode from
   * @param [length] Message length if known beforehand
   * @returns Arrow
   * @throws {Error} If the payload is not a reader or valid buffer
   * @throws {$protobuf.util.ProtocolError} If required fields are missing
   */
  public static decode(
    reader: $protobuf.Reader | Uint8Array,
    length?: number
  ): Arrow

  /**
   * Decodes an Arrow message from the specified reader or buffer, length delimited.
   * @param reader Reader or buffer to decode from
   * @returns Arrow
   * @throws {Error} If the payload is not a reader or valid buffer
   * @throws {$protobuf.util.ProtocolError} If required fields are missing
   */
  public static decodeDelimited(reader: $protobuf.Reader | Uint8Array): Arrow

  /**
   * Verifies an Arrow message.
   * @param message Plain object to verify
   * @returns `null` if valid, otherwise the reason why it is not
   */
  public static verify(message: { [k: string]: any }): string | null

  /**
   * Creates an Arrow message from a plain object. Also converts values to their respective internal types.
   * @param object Plain object
   * @returns Arrow
   */
  public static fromObject(object: { [k: string]: any }): Arrow

  /**
   * Creates a plain object from an Arrow message. Also converts values to other types if specified.
   * @param message Arrow
   * @param [options] Conversion options
   * @returns Plain object
   */
  public static toObject(
    message: Arrow,
    options?: $protobuf.IConversionOptions
  ): { [k: string]: any }

  /**
   * Converts this Arrow to JSON.
   * @returns JSON object
   */
  public toJSON(): { [k: string]: any }

  /**
   * Gets the default type url for Arrow
   * @param [typeUrlPrefix] your custom typeUrlPrefix(default "type.googleapis.com")
   * @returns The default type url
   */
  public static getTypeUrl(typeUrlPrefix?: string): string
}

export namespace Arrow {
  /** EditingMode enum. */
  enum EditingMode {
    READ_ONLY = 0,
    FIXED = 1,
    DYNAMIC = 2,
  }
}
// ...
export interface IArrowNamedDataSet {
  /** ArrowNamedDataSet name */
  name?: string | null

  /** ArrowNamedDataSet hasName */
  hasName?: boolean | null

  /** ArrowNamedDataSet data */
  data?: IArrow | null
}

/** Represents an ArrowNamedDataSet. */
export class ArrowNamedDataSet implements IArrowNamedDataSet {
  /**
   * Constructs a new ArrowNamedDataSet.
   * @param [properties] Properties to set
   */
  constructor(properties?: IArrowNamedDataSet)

  /** ArrowNamedDataSet name. */
  public name: string

  /** ArrowNamedDataSet hasName. */
  public hasName: boolean

  /** ArrowNamedDataSet data. */
  public data?: IArrow | null

  /**
   * Creates a new ArrowNamedDataSet instance using the specified properties.
   * @param [properties] Properties to set
   * @returns ArrowNamedDataSet instance
   */
  public static create(properties?: IArrowNamedDataSet): ArrowNamedDataSet

  /**
   * Encodes the specified ArrowNamedDataSet message. Does not implicitly {@link ArrowNamedDataSet.verify|verify} messages.
   * @param message ArrowNamedDataSet message or plain object to encode
   * @param [writer] Writer to encode to
   * @returns Writer
   */
  public static encode(
    message: IArrowNamedDataSet,
    writer?: $protobuf.Writer
  ): $protobuf.Writer

  /**
   * Encodes the specified ArrowNamedDataSet message, length delimited. Does not implicitly {@link ArrowNamedDataSet.verify|verify} messages.
   * @param message ArrowNamedDataSet message or plain object to encode
   * @param [writer] Writer to encode to
   * @returns Writer
   */
  public static encodeDelimited(
    message: IArrowNamedDataSet,
    writer?: $protobuf.Writer
  ): $protobuf.Writer

  /**
   * Decodes an ArrowNamedDataSet message from the specified reader or buffer.
   * @param reader Reader or buffer to decode from
   * @param [length] Message length if known beforehand
   * @returns ArrowNamedDataSet
   * @throws {Error} If the payload is not a reader or valid buffer
   * @throws {$protobuf.util.ProtocolError} If required fields are missing
   */
  public static decode(
    reader: $protobuf.Reader | Uint8Array,
    length?: number
  ): ArrowNamedDataSet

  /**
   * Decodes an ArrowNamedDataSet message from the specified reader or buffer, length delimited.
   * @param reader Reader or buffer to decode from
   * @returns ArrowNamedDataSet
   * @throws {Error} If the payload is not a reader or valid buffer
   * @throws {$protobuf.util.ProtocolError} If required fields are missing
   */
  public static decodeDelimited(
    reader: $protobuf.Reader | Uint8Array
  ): ArrowNamedDataSet

  /**
   * Verifies an ArrowNamedDataSet message.
   * @param message Plain object to verify
   * @returns `null` if valid, otherwise the reason why it is not
   */
  public static verify(message: { [k: string]: any }): string | null

  /**
   * Creates an ArrowNamedDataSet message from a plain object. Also converts values to their respective internal types.
   * @param object Plain object
   * @returns ArrowNamedDataSet
   */
  public static fromObject(object: { [k: string]: any }): ArrowNamedDataSet

  /**
   * Creates a plain object from an ArrowNamedDataSet message. Also converts values to other types if specified.
   * @param message ArrowNamedDataSet
   * @param [options] Conversion options
   * @returns Plain object
   */
  public static toObject(
    message: ArrowNamedDataSet,
    options?: $protobuf.IConversionOptions
  ): { [k: string]: any }

  /**
   * Converts this ArrowNamedDataSet to JSON.
   * @returns JSON object
   */
  public toJSON(): { [k: string]: any }

  /**
   * Gets the default type url for ArrowNamedDataSet
   * @param [typeUrlPrefix] your custom typeUrlPrefix(default "type.googleapis.com")
   * @returns The default type url
   */
  public static getTypeUrl(typeUrlPrefix?: string): string
}
// ...
export class ArrowVegaLiteChart implements IArrowVegaLiteChart {
  /**
   * Constructs a new ArrowVegaLiteChart.
   * @param [properties] Properties to set
   */
  constructor(properties?: IArrowVegaLiteChart)

  /** ArrowVegaLiteChart spec. */
  public spec: string

  /** ArrowVegaLiteChart data. */
  public data?: IArrow | null

  /** ArrowVegaLiteChart datasets. */
  public datasets: IArrowNamedDataSet[]

  /** ArrowVegaLiteChart useContainerWidth. */
  public useContainerWidth: boolean

  /** ArrowVegaLiteChart theme. */
  public theme: string

  /**
   * Creates a new ArrowVegaLiteChart instance using the specified properties.
   * @param [properties] Properties to set
   * @returns ArrowVegaLiteChart instance
   */
  public static create(properties?: IArrowVegaLiteChart): ArrowVegaLiteChart

  /**
   * Encodes the specified ArrowVegaLiteChart message. Does not implicitly {@link ArrowVegaLiteChart.verify|verify} messages.
   * @param message ArrowVegaLiteChart message or plain object to encode
   * @param [writer] Writer to encode to
   * @returns Writer
   */
  public static encode(
    message: IArrowVegaLiteChart,
    writer?: $protobuf.Writer
  ): $protobuf.Writer

  /**
   * Encodes the specified ArrowVegaLiteChart message, length delimited. Does not implicitly {@link ArrowVegaLiteChart.verify|verify} messages.
   * @param message ArrowVegaLiteChart message or plain object to encode
   * @param [writer] Writer to encode to
   * @returns Writer
   */
  public static encodeDelimited(
    message: IArrowVegaLiteChart,
    writer?: $protobuf.Writer
  ): $protobuf.Writer

  /**
   * Decodes an ArrowVegaLiteChart message from the specified reader or buffer.
   * @param reader Reader or buffer to decode from
   * @param [length] Message length if known beforehand
   * @returns ArrowVegaLiteChart
   * @throws {Error} If the payload is not a reader or valid buffer
   * @throws {$protobuf.util.ProtocolError} If required fields are missing
   */
  public static decode(
    reader: $protobuf.Reader | Uint8Array,
    length?: number
  ): ArrowVegaLiteChart

  /**
   * Decodes an ArrowVegaLiteChart message from the specified reader or buffer, length delimited.
   * @param reader Reader or buffer to decode from
   * @returns ArrowVegaLiteChart
   * @throws {Error} If the payload is not a reader or valid buffer
   * @throws {$protobuf.util.ProtocolError} If required fields are missing
   */
  public static decodeDelimited(
    reader: $protobuf.Reader | Uint8Array
  ): ArrowVegaLiteChart

  /**
   * Verifies an ArrowVegaLiteChart message.
   * @param message Plain object to verify
   * @returns `null` if valid, otherwise the reason why it is not
   */
  public static verify(message: { [k: string]: any }): string | null

  /**
   * Creates an ArrowVegaLiteChart message from a plain object. Also converts values to their respective internal types.
   * @param object Plain object
   * @returns ArrowVegaLiteChart
   */
  public static fromObject(object: { [k: string]: any }): ArrowVegaLiteChart

  /**
   * Creates a plain object from an ArrowVegaLiteChart message. Also converts values to other types if specified.
   * @param message ArrowVegaLiteChart
   * @param [options] Conversion options
   * @returns Plain object
   */
  public static toObject(
    message: ArrowVegaLiteChart,
    options?: $protobuf.IConversionOptions
  ): { [k: string]: any }

  /**
   * Converts this ArrowVegaLiteChart to JSON.
   * @returns JSON object
   */
  public toJSON(): { [k: string]: any }

  /**
   * Gets the default type url for ArrowVegaLiteChart
   * @param [typeUrlPrefix] your custom typeUrlPrefix(default "type.googleapis.com")
   * @returns The default type url
   */
  public static getTypeUrl(typeUrlPrefix?: string): string
}